// src/app/shared/enums.ts
export enum FileType {
    All = 'All Media',
    Audio = 'Audio',
    Document = 'Document',
    Image = 'Image',
    Video = 'Video'
}

export enum SortBy {
    Alphabetical = 'Alphabetical',
    DateUploaded = 'Date Uploaded'
}

export enum SortOrder {
    Ascending = 'Ascending',
    Descending = 'Descending'
}